Разбил файлы на модули и подключил их в main.js
